% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Subgraph_isomorphism.R
\name{Subgraph_isomorphism}
\alias{Subgraph_isomorphism}
\title{This does something}
\usage{
Subgraph_isomorphism(
  pattern.g,
  match.g,
  algorithm = "",
  direction = "",
  graph_type = "",
  constraint = "",
  degsort = 0
)
}
\arguments{
\item{pattern.g}{does something}

\item{match.g}{does something}

\item{algorithm}{does something}

\item{direction}{does something}

\item{graph_type}{does something}

\item{constraint}{does something}

\item{degsort}{does something}
}
\description{
This does something
}
