% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{subgraph_isomorphism_interface}
\alias{subgraph_isomorphism_interface}
\title{does something}
\usage{
subgraph_isomorphism_interface(
  pattern,
  match,
  algorithm,
  direction,
  vertex_atributes_g1,
  vertex_atributes_g2,
  type_graph,
  constraint,
  pattern_vertex,
  match_vertex,
  degreesorted
)
}
\arguments{
\item{pattern}{does something}

\item{match}{does something}

\item{algorithm}{does something}

\item{direction}{does something}

\item{vertex_atributes_g1}{does something}

\item{vertex_atributes_g2}{does something}

\item{type_graph}{does something}

\item{constraint}{does something}

\item{pattern_vertex}{does something}

\item{match_vertex}{does something}

\item{degreesorted}{does something}
}
\description{
does something
}
